Add Database Contract and Database Control
The Database contract now holds a contract for the Resident table.

The Database Control implements the pooling of connection to avoid
the constant creation of connection which may cause performance
issues.

The Database Control will be used to store specialized queries to
better suit the needs of the system while promoting code efficiency.

Also, Resident Control now programmatically generates the labels of
the pages.