Fix label selection limited to 0 - 39 residents.
Wtf. I've been working the whole day creating the resident selection
mechanism and I only just realized that selecting a label from the
resident list paging is limited to only displaying the data of
40 residents, particularly 0 - 39 residents. Created an integer
that will calculate the 'real' index of the resident in the label
selected. Thus, the index is not limited only to the values of 0 and 39.

Also, mResidentSelectedIndex is renamed to mLabelSelectedIndex due
to the fact that it represents the index (0 - 39) of each labels.