General controller to handle image manipulation
WebcamCapture Control only works with 'take photo' feature of the
resident.

Create Photoshop control which will be the combination of
WebcamCaptureControl (Take Picture with webcam)
and SignatureAmendControl (Filters signature to remove background),
and cropping features. Thus, it will serve as the universal
control for all image manipulation.

This controller is still under testing and development.