Use JSON to store preferences rather than Preferences API
The Preferences API is a good candidate to store setting
preferences. However, the data is stored in the registers. Manual
creation of registry keys are needed before the API can properly
function, making it unsuitable for a system that will be handed
to clients with no customer support.

Making use of JSON as a workaround.