Create Barangay ID Report View
JavaFX scene builder is buggy.

The scene created directly from the JavaFX scene builder does not
reflect the actual scene when staged from a java IDE.

Some modifications are made to compensate for the
buggy ui, such as with the TextArea wherein a workaround is needed
to prevent the vertical scroll bar from showing.

Furthermore, an ImageView holding an image outside the project
will not be displayed unless done programmatically.