Put details scenes directly inside information view
It turns out that applying multiple fxml scenes with a single controller
is not recommended, and I see why. The controller has an fx:id that
can only cater to one fxml scene. The controller can only see the
fxml components of one fxml scene and not the other ones.