Display reports of barangay clearance
Barangay clearance reports are now displayed...

~~~~~~~~~~~~~~~~~~~~~~~~~~~ PASSION ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

My grades are faltering due to this project, since I invest most of
my time in the development of this software.

However, I do this not because I'm forced to do it, but rather, it
is of my own accord to do this and I am fully aware of the consequences
of my actions.

A musician doesn't need to master math.
A doctor doesn't need to be an expert at literature.
And entrepreneurship doesn't entirely revolve around computer science.

I've come to realize that the speed of application or implementation of
knowledge is what's important in the world entrepreneuship, and not high
grades.

Thus, from this day onwards I AM AN ENTREPRENEUR.