Finish business clearance form design
Before a business clearance is made, the business itself must be added
in the database.

In addition to generating business clearance reports, this form has the
capability to add the business to the database to allow report generation
for the specific business.

Need to keep on working... cuz self-dictated strict deadline madafaka!