Bug fix in Node Name Handler
The issue was that if there was a single node name, the remove button is
shown due to the rule that every last node name must have a remove button.

Solved the problem by adding another rule that the remove button will only
be shown if there is more than 1 node name visible.