Update BarangayClearanceReportControl class
The information control now asks the barangay ID report to take a
snapshot of the barangay ID to display it
in the details pane.

Documented the whole class and shortened some of the fxml IDs.

Remove the preference model from the Resident Information Form Control,
since the preference model should be handled by the reports to get the
necessary data to generate it.

Note: ResidentInformationFormControl still not completely fixed for
the barangay clearance.