Create business and a Major bug fix
After the creation of a resident or a business, the created entity is
auto selected with an algorithm which includes the computation of the
mCurrentPage.

The problem is, the formula for the mCurrentPage to get to the newly
created entity is flawed, 'Math.ceil(entity index / page capacity - 1)'.

A fucking bug in disguise. Algorithm works well but fails in some situations.

The new formula is 'entity index / page capacity + 1'.