Print mirrored barangay ID
Found a bug with JDK 1.8.
When trying to print the mPrintPane while the said node is rotated in the
x or y axis, a ClassCastException is thrown and the actual printed output
is not equivalent with the expected output. Furthermore, any node that is
rotated within the x or y axis and then printed will yield the same result.

work around to the problem is rotating the mPrintPane by its y axis then
taking a snapshot of the node. The captured snapshot is then plugged image
view to be printed.

The disadvantage of applying the workaround is that the printed output lacks
sharpness, particularly the texts.