Improve JSON file storage system
If the JSON file was manipulated by the user and cannot be parsed,
then delete the JSON file and create it again.

Furthermore, deleting the JSON file function must come after
calling System.gc(). According to the internet, it is another bug.

It came from the internet, so it must be true!